requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.1
djangorestframework==3.11.0
idna==2.8
psycopg2-binary==2.8.4
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
tencentcloud-sdk-python==3.0.115
urllib3==1.25.7
